# internal/trainings/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.3.0 DO NOT EDIT.
package ports

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// Training defines model for Training.
type Training struct {
	CanBeCancelled     bool               `json:"canBeCancelled"`
	MoveProposedBy     *string            `json:"moveProposedBy,omitempty"`
	MoveRequiresAccept bool               `json:"moveRequiresAccept"`
	Notes              string             `json:"notes"`
	ProposedTime       *time.Time         `json:"proposedTime,omitempty"`
	Time               time.Time          `json:"time"`
	User               string             `json:"user"`
	UserUuid           openapi_types.UUID `json:"userUuid"`
	Uuid               openapi_types.UUID `json:"uuid"`
}

// Trainings defines model for Trainings.
type Trainings struct {
	Trainings []Training `json:"trainings"`
}